requirements.txt
python_dotenv==1.0.0
discord.py==2.3.2
tweet-capture==0.2.3
responses==0.24.1
